Refactor: Rearrange structure
