+++ src/components/klassen-verwaltung/DeleteClassDialog.tsx
