+++ src/components/lehrer-verwaltung/DeleteTeacherDialog.tsx
