+++ src/components/lehrer-verwaltung/TeacherAvatar.tsx
