+++ src/components/lehrplan/ClassScheduler.tsx
